pyproject.toml
[tool.poetry]
name = "d11"
version = "0.1.0"
description = ""
authors = ["Alexey Baranov <atlantij@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

aiohttp = { extras = ["speedups"], version = "^3.7.4" }
aiotask-context = "^0.6.1"
beautifulsoup4 = "^4.9.3"
celery = "^5.1.1"
colorlog = "^5.0.1"
dj-database-url = "^0.5.0"
Django = "^3.2.4"
django-admin-rangefilter = "^0.8.1"
django-cacheops = "^6.0"
django-celery-email = "^3.0.0"
django-celery-beat = "^2.2.1"
django-debug-toolbar = "^3.2.1"
django-eml-email-backend = "^0.1"
django-extensions = "^3.1.3"
django-filter = "^2.4.0"
django-json-widget = "^1.1.1"
django-pickling = "^1.0"
django-redis = "^5.0.0"
django-redis-sessions = "^0.6.2"
django-select2 = "^7.7.1"
django-split-settings = "^1.0.1"
djangorestframework = "^3.12.4"
django-prettyjson = "^0.4.1"
django-admin-autocomplete-filter = "^0.7.1"
django-admin-list-filter-dropdown = "^1.0.3"
drf-ujson2 = "^1.6.1"
drf-yasg = "^1.20.0"
django-rest-passwordreset = "^1.2.1"
html5lib = "^1.1"
lxml = "^4.6.3"
Markdown = "^3.3.4"
Pillow = "^8.2.0"
psycopg2-binary = "^2.9.1"
Pygments = "^2.9.0"
python-dateutil = "^2.8.1"
python-json-logger = "^2.0.1"
PyYAML = "^5.4.1"
redis = "^3.5.3"
requests = "^2.25.1"
sorl-thumbnail = "^12.7.0"
ujson = "^4.0.2"
uritemplate = "^3.0.1"
uvloop = "^0.15.2"
uWSGI = "^2.0.19"
watchdog = { extras = ["watchmedo"], version = "^2.1.2" }
flake8 = "^3.9.2"
mypy = "^0.910"
terminaltables = "^3.1.0"
django-webshell = "^0.3.0"
PyMark = "^0.7.1"
pygit2 = "^1.6.1"
python-magic = "^0.4.24"
paramiko = "^2.7.2"
pymarc = "^4.1.1"
marcx = "^0.2.12"
pysmb = "^1.2.7"
sentry-sdk = "^1.4.3"
openpyxl = "^3.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
